backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-mail==0.10.0
python-dotenv==1.0.1